+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Change log
 ## Unreleased
-Nothing yet!
+### Removed
+- Questions can no longer be timed, and the `timeout` setting for quizzes and questions has been removed. (#100)